fixed DI and invalid date bug
