+++ ExerciseTracker.Bata13/STUDY.ConsoleProjects.ExerciseTrackerFour/Service/ExerciseService.cs
@@ -3,7 +3,7 @@ using STUDY.ConsoleProjects.ExerciseTrackerFour.Data.Repository;
 using STUDY.ConsoleProjects.ExerciseTrackerFour.Models;
 
 namespace STUDY.ConsoleProjects.ExerciseTrackerFour.Service;
-internal class ExerciseService : IExerciseService
+public class ExerciseService : IExerciseService
 {
     private readonly IExerciseRepository _exerciseRepository;
     private readonly IUserInput _UserInput;
@@ -12,40 +12,6 @@ internal class ExerciseService : IExerciseService
         _exerciseRepository = exerciseRepository;
         _UserInput = userInput;
     }
-    public void ViewAllExerciseEntries()
-    {
-        var exercises = _exerciseRepository.GetExercises();
-
-        if (exercises.Count == 0)
-        {
-            AnsiConsole.MarkupLine("404 - Cannot be found");
-            MainMenu.ShowMainMenu();
-        }
-
-        else
-        {
-            DataVisualization.ShowDataInTable(exercises);
-        }
-
-        Console.WriteLine("View All Exercise Entries completed");
-    } 
-    public void ViewSpecificExerciseEntry()
-    {        
-        Console.WriteLine("Enter the ID of the exercise entry you want to view:");
-        int exerciseId = int.Parse(Console.ReadLine());        
-
-        Exercise? selectedExercise = _exerciseRepository.GetExerciseEntryById(exerciseId);
-
-        if (selectedExercise is null)
-        {
-            AnsiConsole.MarkupLine("404 - Cannot be found");
-            MainMenu.ShowMainMenu();
-        }              
-
-        Exercise exercise = _exerciseRepository.ViewSpecificExerciseEntry(exerciseId);
-
-        DataVisualization.ShowSingleExercise(exercise);        
-    }    
     public void AddExerciseEntry()
     {
         var (startTime, endTime, duration, comments) = _UserInput.GetUserInputForExcerciseEntry();
