+++ ExerciseTracker.Bata13/STUDY.ConsoleProjects.ExerciseTrackerFour/Service/IExerciseService.cs
@@ -1,8 +1,6 @@
 ﻿namespace STUDY.ConsoleProjects.ExerciseTrackerFour.Service;
 public interface IExerciseService
 {
-    void ViewAllExerciseEntries();
-    void ViewSpecificExerciseEntry();
     void AddExerciseEntry();
     void UpdateExerciseEntry();
     void DeleteExerciseEntry();
